requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
chromadb>=0.4.24
sentence-transformers>=2.6.0
open-clip-torch>=2.24.0
pymupdf>=1.24.0
datasets>=2.18.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
Pillow>=10.2.0
openai>=1.14.0
rouge-score>=0.1.2
bert-score>=0.3.13
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
scikit-learn>=1.4.0
huggingface-hub>=0.21.0
